// frontend/projects/modeling/src/lib/api/generated/dms/model/deltaModification.ts
/**
 * PLASMA DMS Service
 * Service for modeling data sources
 *
 * The version of the OpenAPI document: 1.3.0
 * Contact: pomp@uni-wuppertal.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ObjectProperty } from './objectProperty';
import { SchemaNode } from './schemaNode';
import { SemanticModelNode } from './semanticModelNode';
import { DataProperty } from './dataProperty';
import { Edge } from './edge';


export interface DeltaModification { 
    reference?: string;
    deletion?: boolean;
    entities?: Array<SemanticModelNode>;
    relations?: Array<DataProperty | ObjectProperty>;
    nodes?: Array<SchemaNode>;
    edges?: Array<Edge>;
    anchors?: Array<string>;
    confidence?: number;
}